backend/requirements.txt
flask==2.2.3
pandas
opencv-python
numpy
requests==2.31.0
Flask-Cors==1.10.3
Pillow==8.3.2
scikit-learn==1.3.0